chore(tarot): remove compat shims
