--- apps/tarot/src/avatars/avatar-contracts.ts
@@ -1,1 +1,0 @@
-export type { AvatarsResponse } from '@simpletarot/hooks/server';
