--- apps/tarot/src/avatars/use-rsc-avatar-image.ts
@@ -1,1 +1,0 @@
-export { useRscAvatarImage, type UseRscAvatarImageResult } from '@simpletarot/hooks/client';
