--- apps/tarot/src/readings/reading-contracts.ts
@@ -1,9 +1,0 @@
-export type {
-    ReadingCitation,
-    ReadingHistoryItem,
-    ReadingHistoryResponse,
-    ReadingItem,
-    ReadingPositionResponse,
-    ReadingRequest,
-    ReadingResponse
-} from '@simpletarot/hooks/server';
