--- apps/tarot/src/readings/use-rsc-reading-history.ts
@@ -1,4 +1,0 @@
-export {
-    useRscReadingHistory,
-    type UseRscReadingHistoryResult
-} from '@simpletarot/hooks/client';
